// frontend/src/api/generated/main-service/apiGenerated.ts
export interface DashboardSummaryDto {
  /**
   * @format int32
   * @pattern ^-?(?:0|[1-9]\d*)$
   */
  totalGenerated: number | string
  totalGeneratedTrend: (number | string)[]
  /**
   * @format int32
   * @pattern ^-?(?:0|[1-9]\d*)$
   */
  activeProjects: number | string
  /**
   * @format int32
   * @pattern ^-?(?:0|[1-9]\d*)$
   */
  maxProjects: number | string
  /**
   * @format double
   * @pattern ^-?(?:0|[1-9]\d*)(?:\.\d+)?(?:[eE][+-]?\d+)?$
   */
  successRate: number | string
  successRateTrend: (number | string)[]
  /**
   * @format int32
   * @pattern ^-?(?:0|[1-9]\d*)$
   */
  totalTemplates: number | string
  templateBreakdown: TemplateBreakdownDto[]
  /**
   * @format int32
   * @pattern ^-?(?:0|[1-9]\d*)$
   */
  creditBalance: number | string
  /**
   * @format int32
   * @pattern ^-?(?:0|[1-9]\d*)$
   */
  creditLimit: number | string
}
// ...
export interface GenerationDto {
  /** @format uuid */
  id: string
  templateName: null | string
  templateKey: null | string
  type: null | string
  status: string
  /** @format date-time */
  createdDatetime: any
  /**
   * @format int64
   * @pattern ^-?(?:0|[1-9]\d*)$
   */
  durationMs: null | number | string
  /**
   * @format int64
   * @pattern ^-?(?:0|[1-9]\d*)$
   */
  fileSizeBytes: null | number | string
  errorMessage: null | string
  downloadUrl: null | string
}
// ...
export interface TemplateBreakdownDto {
  type: string
  /**
   * @format int32
   * @pattern ^-?(?:0|[1-9]\d*)$
   */
  count: number | string
}
// ...
export interface TemplatePerformanceDto {
  templateKey: string
  templateName: string
  projectName: string
  type: string
  /**
   * @format int32
   * @pattern ^-?(?:0|[1-9]\d*)$
   */
  totalGenerations: number | string
  /**
   * @format double
   * @pattern ^-?(?:0|[1-9]\d*)(?:\.\d+)?(?:[eE][+-]?\d+)?$
   */
  avgDurationMs: number | string
  /**
   * @format int64
   * @pattern ^-?(?:0|[1-9]\d*)$
   */
  avgFileSizeBytes: number | string
  /**
   * @format int32
   * @pattern ^-?(?:0|[1-9]\d*)$
   */
  errorCount: number | string
  /**
   * @format double
   * @pattern ^-?(?:0|[1-9]\d*)(?:\.\d+)?(?:[eE][+-]?\d+)?$
   */
  successRate: number | string
  /**
   * @format double
   * @pattern ^-?(?:0|[1-9]\d*)(?:\.\d+)?(?:[eE][+-]?\d+)?$
   */
  errorRate: number | string
  /** @format date-time */
  lastGeneratedAt: any
}
// ...
export interface UsageDataDto {
  range: string
  groupBy: string
  /**
   * @format int32
   * @pattern ^-?(?:0|[1-9]\d*)$
   */
  totalVolume: number | string
  data: UsageDataPointDto[]
}

export interface UsageDataPointDto {
  date: string
  /**
   * @format int32
   * @pattern ^-?(?:0|[1-9]\d*)$
   */
  count: number | string
  breakdown: Record<string, number | string>
}